lib-sasl: Fix oauthbearer compatibility with 2.2

+++ src/lib-sasl/mech-oauthbearer.c
@@ -17,7 +17,7 @@ struct oauthbearer_dsasl_client {
 
 static int
 mech_oauthbearer_input(struct dsasl_client *_client,
-		 const unsigned char *input, size_t input_len,
+		 const unsigned char *input, unsigned int input_len,
 		 const char **error_r)
 {
 	struct oauthbearer_dsasl_client *client =
@@ -78,7 +78,7 @@ mech_oauthbearer_input(struct dsasl_client *_client,
 
 static int
 mech_oauthbearer_output(struct dsasl_client *_client,
-		  const unsigned char **output_r, size_t *output_len_r,
+		  const unsigned char **output_r, unsigned int *output_len_r,
 		  const char **error_r)
 {
 	struct oauthbearer_dsasl_client *client =
@@ -112,7 +112,7 @@ mech_oauthbearer_output(struct dsasl_client *_client,
 
 static int
 mech_xoauth2_output(struct dsasl_client *_client,
-		    const unsigned char **output_r, size_t *output_len_r,
+		    const unsigned char **output_r, unsigned int *output_len_r,
 		    const char **error_r)
 {
 	struct oauthbearer_dsasl_client *client =
